refractored

+++ CP/Codeforces/Div2 B/Tprime.cpp
@@ -3,12 +3,13 @@ using namespace std;
 
 
 bool isTprime(long long int a){
-    if(sqrt(a)*sqrt(a)!=a)
+    if((long long int)sqrt(a)*(long long int)sqrt(a)!=a)
     return false;
     
     if (a <= 1)
         return false;
-    for (long long int i = 2; i <sqrt(a); i++)
+
+    for (long long int i = 2; i < (long long int)sqrt(a); i++)
         if (a % i == 0)
             return false;
  
@@ -16,7 +17,7 @@ bool isTprime(long long int a){
 }
 int main(){
 
-    int n; cin>>n;
+    long long int n; cin>>n;
     vector<long long int> arr(n);
     for(long long int i=0;i<n;i++)
     cin>>arr[i];
